storage: use connection string for postgres
